+++ src/clj_service/pedestal/interceptors/auth.clj
@@ -1,18 +1,12 @@
-(ns common-labsoft.pedestal.interceptors.auth
+(ns clj-service.pedestal.interceptors.auth
   (:require [clojure.string :as str]
-            [common-labsoft.protocols.token :as protocols.token]
             [io.pedestal.log :as log]
-            [common-labsoft.datomic.api]
-            [common-labsoft.exception :as exception]
-            [datomic.api :as d]))
+            [clj-service.exception :as exception]))
 
 (defn headers->token [headers]
   (some-> headers (get "authorization") (str/split #"Bearer ") second))
 
 (defn allowed-scope? [scopes required] (some #((set required) %) scopes))
-
-(defn token-component [context] (-> context :request :components :token))
-(defn datomic-component [context] (-> context :request :components :datomic))
 
 (defn forbidden []
   (log/error :log :forbidden-access)
@@ -31,28 +25,10 @@
         auth-id (str (get-in context [:request :identity :token/sub]))]
     (= request-id auth-id)))
 
-(defn auth-owner? [res-key query context]
-  (let [auth-id (get-in context [:request :identity :token/sub])
-        res-id (get-in context [:request :path-params res-key])
-        datomic (datomic-component context)]
-    (d/q {:find  '[?entity .]
-          :in    '[$ ?owner-id ?resource-id]
-          :where query}
-         (common-labsoft.datomic.api/db datomic)
-         auth-id
-         res-id)))
-
-(defn verify-token [context]
-  (some->> context
-           :request
-           :headers
-           headers->token
-           (protocols.token/decode (token-component context))))
-
-(def auth
+(defn auth [verify-fn]
   {:name  ::auth
    :enter (fn [context]
-            (if-let [claim (verify-token context)]
+            (if-let [claim (verify-fn context)]
               (assoc-in context [:request :identity] claim)
               context))})
 
@@ -63,10 +39,6 @@
 (defn allow-types? [& types]
   {:name  ::allow-types?
    :enter (fn [context] (if (auth-type? types context) context (forbidden)))})
-
-(defn owner? [res-key query]
-  {:name  ::owner?
-   :enter (fn [context] (if (auth-owner? res-key query context) context (forbidden)))})
 
 (defn user-identity [path-id]
   {:name  :user-identity