+++ src/clj_service/test_helpers.clj
@@ -1,17 +1,11 @@
-(ns common-labsoft.test-helpers
-  (:require [common-labsoft.exception :as exception]
-            [common-labsoft.protocols.config :as protocols.config]
-            [common-labsoft.datomic.api :as datomic]
+(ns clj-service.test-helpers
+  (:require [clj-service.exception :as exception]
+            [clj-service.protocols.config :as protocols.config]
             [io.pedestal.test :refer [response-for]]
             [cheshire.core :as chesire]
-            [com.stuartsierra.component :as component]
             [io.pedestal.http :as http]
-            [datomic.api :as d]
-            [common-labsoft.pedestal.interceptors.auth]
-            [common-labsoft.datomic.transform :as datomic.transform]
-            [common-labsoft.components.datomic :as components.datomic]
-            [common-labsoft.time :as time]
-            [common-labsoft.misc :as misc]
+            [clj-service.pedestal.interceptors.auth]
+            [clj-service.time :as time]
             [cheshire.core :as cheshire]))
 
 (defonce ^:private service-fn (atom nil))
@@ -24,43 +18,11 @@
     (get-in-maybe [this ks] (get-in obj ks))
     (get-in! [this ks] (or (get-in obj ks) (exception/not-found! {})))))
 
-(defn db-with-entities
-  "Return a db with seed tx-data applied"
-  ([dtm entities]
-   (-> (datomic.api/db dtm)
-       (d/with (map datomic.transform/transform-to-datomic entities))
-       :db-after)))
-
 (defmacro as-of [time & body]
-  `(with-redefs [time/now   (fn [] ~time)
+  `(with-redefs [time/now (fn [] ~time)
                  time/today (fn [] ~time)]
      (do
        ~@body)))
-
-(defn clean-datomic [dtm]
-  (d/delete-database (:endpoint dtm))
-  (component/stop dtm)
-  (component/start dtm))
-
-(defn transact-entities [datomic entities]
-  @(d/transact (datomic/conn datomic) (map datomic.transform/transform-to-datomic entities))
-  datomic)
-
-(defn new-mock-datomic [datomic-settings]
-  (let [config (mock-config {:datomic-endpoint (str "datomic:mem://" (misc/uuid))})]
-    (component/start (components.datomic/map->Datomic {:settings datomic-settings
-                                                       :config   config}))))
-
-(defmacro with-entities [datomic-settings [datomic-binding db-binding] entities & body]
-  `(do
-     (let [~datomic-binding (new-mock-datomic ~datomic-settings)
-           _# (transact-entities ~datomic-binding ~entities)
-           ~db-binding (datomic/db ~datomic-binding)
-           result# (do ~@body)]
-       (d/delete-database (:endpoint ~datomic-binding))
-       (component/stop ~datomic-binding)
-       (d/release (datomic/conn ~datomic-binding))
-       result#)))
 
 (defn test-service
   "Return a service-fn for use with Pedestal's `response-for` test helper."
@@ -76,11 +38,6 @@
        result#)
      (finally
        (~stop-fn))))
-
-(defmacro with-token [token & body]
-  `(with-redefs [common-labsoft.pedestal.interceptors.auth/verify-token (constantly ~token)]
-     (do
-       ~@body)))
 
 (defn get-pedestal-service [system]
   (::http/service-fn (-> system :pedestal :service)))