+++ src/common_labsoft/protocols/crypto.clj
@@ -1,7 +1,0 @@
-(ns common-labsoft.protocols.crypto)
-
-(defprotocol Crypto
-  (bcrypt [this value])
-  (check  [this encrypted attempt]))
-
-(def ICrypto (:on-interface Crypto))