+++ src/common_labsoft/protocols/datomic.clj
@@ -1,7 +1,0 @@
-(ns common-labsoft.protocols.datomic)
-
-(defprotocol Datomic
-  (connection [this])
-  (db [this]))
-
-(def IDatomic (:on-interface Datomic))