+++ src/common_labsoft/protocols/token.clj
@@ -1,8 +1,0 @@
-(ns common-labsoft.protocols.token)
-
-(defprotocol Token
-  (encode [this content])
-  (decode [this token])
-  (verify [this token]))
-
-(def IToken (:on-interface Token))