+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Several improvements to time and datomic schema handling
 
 ### [0.5.0-SNAPSHOT]
-- Move system definition to outside of common-labsoft
+- Move system definition to outside of clj-service
 
 ### [0.4.1-SNAPSHOT]
 - Improvements and small fixes